requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==2.4.1
bleach==6.2.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
dnspython==2.8.0
elevenlabs==2.35.0
email-validator==2.3.0
executing==2.0.1
fastapi==0.129.0
fastar==0.8.0
google-generativeai>=0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.2
jupyter_core==5.7.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.3.1
packaging==24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==6.30.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.4.0
Pygments==2.17.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.22
python-slugify==8.0.4
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==26.0.0
requests==2.32.4
rignore==0.7.6
scikit-learn==1.5.1
scipy==1.15.2
sentry-sdk==2.52.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==0.52.1
text-unidecode==1.3
threadpoolctl==3.6.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==16.0
Werkzeug==3.1.3
